requirements.txt
cryptography>=42.0.5
rich>=13.7.0
PySide6>=6.7.2
pytest>=8.3.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
requests>=2.31.0
PyJWT>=2.8.0

